requirements.txt
# Telegram Bot - Restaurant Dish Search
aiogram>=3.4.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Agent (browser-based menu finder - fallback)
playwright>=1.40.0
